Fix read/write support for single-parity fields, preserving prior default behaviour for Nc=3 fields which does promotion to dual parity.  This fix allows us to read/write single-parity coarse grid eigen-vectors correctly between node geometry changes
