+++ include/layout_hyper.h
@@ -1,13 +1,11 @@
-#ifndef LAYOUT_HYPER_H
-#define LAYOUT_HYPER_H
+#pragma once
 
 #ifdef __cplusplus
 extern "C" {
 #endif
-  /* layout_hyper */
   /* These routines get a quda_* prefix to avoid
-     linker conflicts with MILC */
-  int quda_setup_layout(int len[], int nd, int numnodes);
+     potential linker conflicts, with MILC */
+  int quda_setup_layout(int len[], int nd, int numnodes, int single_parity);
   int quda_node_number(const int x[]);
   int quda_node_index(const int x[]);
   void quda_get_coords(int x[], int node, int index);
@@ -17,5 +15,3 @@ extern "C" {
 #ifdef __cplusplus
 }
 #endif
-
-#endif //  LAYOUT_HYPER_H
